Barely working

+++ srcs/wordpress/srcs/entry.ash
@@ -1,20 +1,29 @@
 #!/bin/ash
 
-curl -LO https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
-mv wp-cli.phar /usr/local/bin/wp
-chmod +x /usr/local/bin/wp
-/usr/local/bin/wp core download --allow-root --path=/var/www/html
-mv /var/www/html/wp-config-sample.php /var/www/html/wp-config.php
-sed -i -e "s/database_name_here/wordpress/g" -e "s/username_here/user1/g" -e "s/password_here/user123/g" -e "s/localhost/mariadb/g" /var/www/html/wp-config.php
+if [ -e "/usr/local/bin/wp" ]; then
+	echo "WPCLI ready"
+else {
+	curl -LO https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+	mv wp-cli.phar /usr/local/bin/wp
+	chmod +x /usr/local/bin/wp
+	/usr/local/bin/wp core download --allow-root --path=/var/www/html
+}
+fi
+
+if [ -e "/var/www/html/wp-config.php" ]; then
+{
+	sed -i -e "s/database_name_here/wordpress/g" -e "s/username_here/user1/g" -e "s/password_here/user123/g" -e "s/localhost/mariadb/g" /var/www/html/wp-config.php
+	echo "DB UPDATED"
+}
+else
+{
+	mv /var/www/html/wp-config-sample.php /var/www/html/wp-config.php
+	sed -i -e "s/database_name_here/wordpress/g" -e "s/username_here/user1/g" -e "s/password_here/user123/g" -e "s/localhost/mariadb/g" /var/www/html/wp-config.php
+}
+fi
+
 /usr/local/bin/wp core install --allow-root --path=/var/www/html --url=ahkecha.42.fr \
-	--admin_user=admin --admin_password=admin --admin_email=admin@admin.com --title="iwakgjdzgjkadshj"
-# if grep -Fxq "listen = 0.0.0.0:9000" /etc/php7/php-fpm.d/www.conf
-# then
-# 	echo "www.conf OK"
-# else
-# 	# sed -i -e 's/^listen.*=.*\/listen = 127.0.0.1:9000/' /etc/php8/php-fpm.d/www.conf
-# 	echo "listen = 127.0.0.1:9000" >> /etc/php7/php-fpm.d/www.conf
-# 	echo "OK"
-# fi
+	--admin_user=admin --admin_password=admin --admin_email=admin@admin.com --title="Wordpress Website test"
+
 echo "Running php-fpm"
 php-fpm7 -F -R
